requirements.txt
# Core dependencies (compatible versions for Python 3.13)
chromadb>=0.4.20,<0.5.0
requests>=2.31.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Optional but recommended
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.65.0,<5.0.0
rich>=13.0.0,<14.0.0

# Development dependencies
pytest>=7.4.0,<8.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
jupyter>=1.0.0,<2.0.0

# For future enhancements
streamlit>=1.28.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0